Notifications Design İmprovement
